requirements.txt
asgiref==3.4.1
Django==3.2.7
django-environ==0.7.0
djangorestframework==3.12.4
mysqlclient==2.0.3
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.2
